requirements.txt
# Azure Government Flask RAG Application
# Core dependencies
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0

# Semantic Kernel
semantic-kernel>=1.15.0

# Azure SDKs
azure-identity>=1.19.0
azure-cosmos>=4.7.0
azure-search-documents>=11.4.0

# OpenAI (for embeddings if needed)
openai>=1.50.0

# Async support
aiohttp>=3.9.0
aiofiles>=23.0.0

# Utilities
pydantic>=2.0.0
